requirements.txt
streamlit
sounddevice
scipy
whisper
python-dotenv
requests
soundfile
numpy
faster-whisper
transformers
torch
huggingface-hub
sentence-transformers